Initial commit for simple_spheres--a very basic, fairly fast multithreaded raytracer written in C.
I'm using it mainly as a demo and testbed.

It can render the canonical shiny balls -- eight of them,
bouncing around in an invisible box, with two point
light sources and a trace depth of two (ie you can see
reflections of reflections) -- at 30fps in my little VM
running on my laptop.

It is fairly badass.