Added a very similar but even more simplified raytracer in Haskell.
Pros:
The entire rendering is a single pure function, which is pretty cool.
In theory this makes the individual ray casts trivially parallelizable.

Cons:
Even with the cool parts like antialiasing, collision detection, and
multiple lightsources disabled and the trace depth reduced to 1 at
320x240 (ie extremely limited settings), roughly 1 FPS.

The C code can raytrace a 640x480 image with all the cool stuff
mentioned above, and still get 50fps on an old-ish Thinkpad.